requirements/requirements.txt
asgiref==3.7.2
distlib==0.3.8
Django==5.0.1
django-ckeditor==6.7.0
django-filter==23.5
django-js-asset==2.2.0
djangorestframework==3.14.0
filelock==3.13.1
Markdown==3.5.2
platformdirs==4.1.0
psycopg2-binary==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
virtualenv==20.25.0
